+++ packages/agent-app-sdk/src/widgets/details/details-widget.ts
@@ -1,9 +1,21 @@
-import { createConnection, createWidget, IConnection, withAmplitude, withPayments, withToasts } from '@livechat/widget-core-sdk';
+import {
+  createConnection,
+  createWidget,
+  IConnection,
+  withAmplitude,
+  withChatWidget,
+  withPayments,
+  withToasts
+} from '@livechat/widget-core-sdk';
 import { withCustomerProfile } from '../shared/customer-profile';
 import { withRichMessages } from '../shared/rich-messages';
 import { withTheme } from '../shared/theme';
 import assertSection from './custom-sections';
-import { IDetailsWidgetApi, IDetailsWidgetEvents, ISection } from './interfaces';
+import {
+  IDetailsWidgetApi,
+  IDetailsWidgetEvents,
+  ISection
+} from './interfaces';
 
 export function DetailsWidget(connection: IConnection<IDetailsWidgetEvents>) {
   const base = createWidget<IDetailsWidgetApi, IDetailsWidgetEvents>(
@@ -25,7 +37,15 @@ export function DetailsWidget(connection: IConnection<IDetailsWidgetEvents>) {
     }
   );
 
-  const widget = withAmplitude(withRichMessages(withCustomerProfile(withTheme(withToasts(withPayments(base))))));
+  const widget = withAmplitude(
+    withRichMessages(
+      withCustomerProfile(
+        withTheme(
+          withChatWidget(withChatWidget(withToasts(withPayments(base))))
+        )
+      )
+    )
+  );
 
   return widget;
 }