+++ packages/agent-app-sdk/src/widgets/settings/settings-widget.ts
@@ -1,4 +1,12 @@
-import { createConnection, createWidget, IConnection, withAmplitude, withPayments, withToasts } from '@livechat/widget-core-sdk';
+import {
+  createConnection,
+  createWidget,
+  IConnection,
+  withAmplitude,
+  withChatWidget,
+  withPayments,
+  withToasts
+} from '@livechat/widget-core-sdk';
 import { withPageData } from '../shared/page-data';
 import { withTheme } from '../shared/theme';
 import { ISettingsWidgetApi, ISettingsWidgetEvents } from './interfaces';
@@ -12,7 +20,9 @@ export function SettingsWidget(connection: IConnection<ISettingsWidgetEvents>) {
       }
     }
   );
-  return withAmplitude(withPageData(withTheme(withToasts(withPayments(base)))));
+  return withAmplitude(
+    withPageData(withTheme(withChatWidget(withToasts(withPayments(base)))))
+  );
 }
 
 export type ISettingsWidget = ReturnType<typeof SettingsWidget>;