+++ packages/helpdesk-sdk/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @livechat/helpdesk-sdk
 
+## 1.4.0
+
+### Minor Changes
+
+- fd24557: Add functionality to open Chat Widget in the parent environment
+
+### Patch Changes
+
+- Updated dependencies [fd24557]
+  - @livechat/widget-core-sdk@1.4.0
+
 ## 1.3.0
 
 ### Minor Changes